Move output to its own file
